Changed convertSize (int) to convertSize( unsigned long ) to accomodate very large sized files.  Also added i18n support. (Dawit A.)
svn path=/trunk/kdelibs/; revision=33541

+++ kio/kio_job.h
@@ -235,7 +235,7 @@ public:
    * @param  _size  size in bytes
    * @return  converted size as a string - e.g. 123.4 kB , 12 MB
    */
-  static QString convertSize( int size );
+  static QString convertSize( unsigned long size );
 
   friend KIOListViewItem;
